annotation/requirements.txt
pyspark
aiosqlite==0.18.0
fastapi
uvicorn
Pillow
numpy
kmeans
scikit-learn
matplotlib
scipy
webcolors
pandas
requests